requirements.txt
PyQt5
mss
openai
Pillow
pyautogui
